move filter specs to filters module
