+++ corehq/apps/userreports/reports/specs.py
@@ -19,9 +19,6 @@ from sqlagg.columns import (
     YearColumn,
 )
 from corehq.apps.reports.sqlreport import DatabaseColumn, AggregateColumn
-from corehq.apps.userreports.indicators.specs import DataTypeProperty
-from corehq.apps.userreports.reports.filters import DateFilterValue, ChoiceListFilterValue, \
-    NumericFilterValue
 from corehq.apps.userreports.specs import TypeProperty
 from corehq.apps.userreports.sql import get_expanded_column_config, SqlColumnConfig
 from corehq.apps.userreports.transforms.factory import TransformFactory
@@ -39,22 +36,6 @@ SQLAGG_COLUMN_MAP = {
     'simple': SimpleColumn,
     'year': YearColumn,
 }
-
-
-class ReportFilter(JsonObject):
-    type = StringProperty(required=True)
-    slug = StringProperty(required=True)
-    field = StringProperty(required=True)
-    display = DefaultProperty()
-    compare_as_string = BooleanProperty(default=False)
-
-    def create_filter_value(self, value):
-        return {
-            'date': DateFilterValue,
-            'numeric': NumericFilterValue,
-            'choice_list': ChoiceListFilterValue,
-            'dynamic_choice_list': ChoiceListFilterValue,
-        }[self.type](self, value)
 
 
 class ReportColumn(JsonObject):
@@ -291,64 +272,6 @@ class PercentageColumn(ReportColumn):
 def _add_column_id_if_missing(obj):
     if obj.get('column_id') is None:
         obj['column_id'] = obj.get('alias') or obj['field']
-
-
-class FilterChoice(JsonObject):
-    value = DefaultProperty()
-    display = StringProperty()
-
-    def get_display(self):
-        return self.display or self.value
-
-
-def _validate_filter_slug(s):
-    if "-" in s:
-        raise Exception(_(
-            """
-            Filter slugs must be legal sqlalchemy bind parameter names.
-            '-' and other special character are prohibited
-            """
-        ))
-
-
-class FilterSpec(JsonObject):
-    """
-    This is the spec for a report filter - a thing that should show up as a UI filter element
-    in a report (like a date picker or a select list).
-    """
-    type = StringProperty(required=True, choices=['date', 'numeric', 'choice_list', 'dynamic_choice_list'])
-    slug = StringProperty(required=True, validators=_validate_filter_slug)  # this shows up as the ID in the filter HTML.
-    field = StringProperty(required=True)  # this is the actual column that is queried
-    display = DefaultProperty()
-    datatype = DataTypeProperty(default='string')
-
-    def get_display(self):
-        return self.display or self.slug
-
-
-class DateFilterSpec(FilterSpec):
-    compare_as_string = BooleanProperty(default=False)
-
-
-class ChoiceListFilterSpec(FilterSpec):
-    type = TypeProperty('choice_list')
-    show_all = BooleanProperty(default=True)
-    datatype = DataTypeProperty(default='string')
-    choices = ListProperty(FilterChoice)
-
-
-class DynamicChoiceListFilterSpec(FilterSpec):
-    type = TypeProperty('dynamic_choice_list')
-    show_all = BooleanProperty(default=True)
-    datatype = DataTypeProperty(default='string')
-
-    @property
-    def choices(self):
-        return []
-
-
-class NumericFilterSpec(FilterSpec):
-    type = TypeProperty('numeric')
 
 
 class ChartSpec(JsonObject):
